# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: profiles
    description: User profile operations
  - name: serviceReports
    description: Service report operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /profile:
    get:
      operationId: getCurrentProfile
      tags: [profiles]
      summary: Get current user profile
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateCurrentProfile
      tags: [profiles]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /service-reports:
    get:
      operationId: listServiceReports
      tags: [serviceReports]
      summary: List service reports visible to the current user
      responses:
        "200":
          description: Service reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceReport"
        "401":
          description: Unauthorized
    post:
      operationId: createServiceReport
      tags: [serviceReports]
      summary: Create a service report for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceReportRequest"
      responses:
        "201":
          description: Created report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceReport"
        "401":
          description: Unauthorized
  /service-reports/{id}:
    patch:
      operationId: updateServiceReport
      tags: [serviceReports]
      summary: Update review status and notes, supervisor only
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceReportRequest"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceReport"
        "401":
          description: Unauthorized
        "403":
          description: Supervisor role required
        "404":
          description: Report not found
  /service-reports/summary:
    get:
      operationId: getServiceReportsSummary
      tags: [serviceReports]
      summary: Dashboard totals for reports visible to the current user
      responses:
        "200":
          description: Service report summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceReportsSummary"
        "401":
          description: Unauthorized
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    UserRole:
      type: string
      enum:
        - technician
        - supervisor
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        displayName:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - displayName
        - email
        - role
        - createdAt
        - updatedAt
    UpdateUserProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
        role:
          $ref: "#/components/schemas/UserRole"
      minProperties: 1
    ServiceReport:
      type: object
      properties:
        id:
          type: integer
        ownerUserId:
          type: string
        ownerName:
          type: string
        ownerEmail:
          type: string
        technicianName:
          type: string
        workDate:
          type: string
          format: date
        shiftLabel:
          type: string
        serviceActivity:
          type: string
        overtime50Normal:
          type: number
        overtime50NormalKm40:
          type: number
        overtime50WeekendHoliday:
          type: number
        overtime50WeekendHolidayKm40:
          type: number
        overtime100Normal:
          type: number
        overtime100NormalKm40:
          type: number
        overtime100WeekendHoliday:
          type: number
        overtime100WeekendHolidayKm40:
          type: number
        soloKm40:
          type: boolean
        soloKm40Hours:
          type: number
        technicalAssistanceGuard:
          type: number
        fieldActivation:
          type: number
        total50Hours:
          type: number
        total100Hours:
          type: number
        totalKm40Items:
          type: integer
        totalAdditionalItems:
          type: number
        reviewed:
          type: boolean
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ownerUserId
        - ownerName
        - ownerEmail
        - technicianName
        - workDate
        - shiftLabel
        - serviceActivity
        - overtime50Normal
        - overtime50NormalKm40
        - overtime50WeekendHoliday
        - overtime50WeekendHolidayKm40
        - overtime100Normal
        - overtime100NormalKm40
        - overtime100WeekendHoliday
        - overtime100WeekendHolidayKm40
        - soloKm40
        - soloKm40Hours
        - technicalAssistanceGuard
        - fieldActivation
        - total50Hours
        - total100Hours
        - totalKm40Items
        - totalAdditionalItems
        - reviewed
        - notes
        - createdAt
    CreateServiceReportRequest:
      type: object
      properties:
        technicianName:
          type: string
          minLength: 1
        workDate:
          type: string
          format: date
        shiftLabel:
          type: string
        serviceActivity:
          type: string
          minLength: 1
        overtime50Normal:
          type: number
          minimum: 0
        overtime50NormalKm40:
          type: number
          minimum: 0
        overtime50WeekendHoliday:
          type: number
          minimum: 0
        overtime50WeekendHolidayKm40:
          type: number
          minimum: 0
        overtime100Normal:
          type: number
          minimum: 0
        overtime100NormalKm40:
          type: number
          minimum: 0
        overtime100WeekendHoliday:
          type: number
          minimum: 0
        overtime100WeekendHolidayKm40:
          type: number
          minimum: 0
        soloKm40:
          type: boolean
        soloKm40Hours:
          type: number
          minimum: 0
        technicalAssistanceGuard:
          type: number
          minimum: 0
        fieldActivation:
          type: number
          minimum: 0
        notes:
          type: string
      required:
        - technicianName
        - workDate
        - serviceActivity
        - overtime50Normal
        - overtime50NormalKm40
        - overtime50WeekendHoliday
        - overtime50WeekendHolidayKm40
        - overtime100Normal
        - overtime100NormalKm40
        - overtime100WeekendHoliday
        - overtime100WeekendHolidayKm40
        - soloKm40
        - soloKm40Hours
        - technicalAssistanceGuard
        - fieldActivation
    UpdateServiceReportRequest:
      type: object
      properties:
        reviewed:
          type: boolean
        notes:
          type: string
      minProperties: 1
    ServiceReportsSummary:
      type: object
      properties:
        totalReports:
          type: integer
        pendingReview:
          type: integer
        reviewedReports:
          type: integer
        total50Hours:
          type: number
        total100Hours:
          type: number
        totalKm40Items:
          type: integer
        totalSoloKm40Hours:
          type: number
        totalGuardias:
          type: number
        totalActivaciones:
          type: number
        latestReportDate:
          type: string
          nullable: true
      required:
        - totalReports
        - pendingReview
        - reviewedReports
        - total50Hours
        - total100Hours
        - totalKm40Items
        - totalSoloKm40Hours
        - totalGuardias
        - totalActivaciones
        - latestReportDate