# lib/api-zod/src/generated/types/serviceReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceReport {
  id: number;
  ownerUserId: string;
  ownerName: string;
  ownerEmail: string;
  technicianName: string;
  workDate: Date;
  shiftLabel: string;
  serviceActivity: string;
  overtime50Normal: number;
  overtime50NormalKm40: number;
  overtime50WeekendHoliday: number;
  overtime50WeekendHolidayKm40: number;
  overtime100Normal: number;
  overtime100NormalKm40: number;
  overtime100WeekendHoliday: number;
  overtime100WeekendHolidayKm40: number;
  soloKm40: boolean;
  soloKm40Hours: number;
  technicalAssistanceGuard: number;
  fieldActivation: number;
  total50Hours: number;
  total100Hours: number;
  totalKm40Items: number;
  totalAdditionalItems: number;
  reviewed: boolean;
  notes: string;
  createdAt: Date;
}