# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface UserProfile {
  id: number;
  userId: string;
  displayName: string;
  email: string;
  role: UserRole;
  createdAt: Date;
  updatedAt: Date;
}